backend/requirements.txt
# Core Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
alembic==1.13.1

# AI/ML
anthropic==0.18.0
openai==1.9.0  # Still needed for embeddings (text-embedding-3-small)
langchain==0.1.0
langchain-anthropic==0.1.0
pillow==10.2.0
numpy==1.26.3
scikit-learn==1.4.0
sentence-transformers==2.2.2  # Alternative embedding option

# Web Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
redis==5.0.1
celery==5.3.4

# Development
pytest==7.4.4
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# WebSocket support
websockets==12.0
python-socketio==5.11.0